Setup migration for all notes to corresond to the model
the reason to use very_long is that it makes more sense
than an arbitary long length that supports the current
use case. the very long is based on when data can be
stored in row, so has meaning to the database and will
not change (unless db settings are changed).