Use _id instead, as `id` is subject to change
The `id` of a dataset can change when a user changes its name or
the datasets' name. The `_id` is persistent, which means that we
can keep track of whether or not the dataset has already been
indexed. If any part of the name changes, the old `name` simply
redirects. In the worst case (bare _id becoming invalid), we could
always reindex and match based on `_id`.